+++ NAStify/ConnectionManager/Plugins/Dropbox/CMDropbox.h
@@ -61,6 +61,6 @@
 - (NetworkConnection *)urlForFile:(FileItem *)file;
 
 /* Server features */
-- (NSInteger)supportedFeaturesAtPath:(NSString *)path;
+- (long long)supportedFeaturesAtPath:(NSString *)path;
 - (NSInteger)supportedSharingFeatures;
 @end
