+++ NAStify/ConnectionManager/Plugins/Dropbox/CMDropbox.m
@@ -613,7 +613,7 @@
 
 #pragma mark - supported features
 
-- (NSInteger)supportedFeaturesAtPath:(NSString *)path
+- (long long)supportedFeaturesAtPath:(NSString *)path
 {
     return (CMSupportedFeaturesMaskFileDelete      |
             CMSupportedFeaturesMaskFolderDelete    |
