+++ NAStify/ConnectionManager/Plugins/Mega/CMMega.m
@@ -253,9 +253,9 @@
 
 #pragma mark - supported features
 
-- (NSInteger)supportedFeaturesAtPath:(NSString *)path
+- (long long)supportedFeaturesAtPath:(NSString *)path
 {
-    NSInteger features = CMSupportedFeaturesMaskFolderCreate   |
+    long long features = CMSupportedFeaturesMaskFolderCreate   |
                          CMSupportedFeaturesMaskFileDelete     |
                          CMSupportedFeaturesMaskFolderDelete   |
                          CMSupportedFeaturesMaskDeleteCancel   |
